post command query completed

+++ Application/Common/Models/UserDto.cs
@@ -14,6 +14,6 @@
 
         public string? Email { get; set; }
 
-        public short RoleId { get; set; }
+        public int RoleId { get; set; }
     }
 }
