+++ Domain/Entities/User.cs
@@ -16,7 +16,7 @@ namespace Domain.Entities
 
         public string? Email { get; set; }
 
-        public short RoleId { get; set; }
+        public int RoleId { get; set; }
 
         public Role? Role { get; set; }
